Move and refactor Coordinates.java from person to cell
